ai-service/main.py
from fastapi import FastAPI, UploadFile, File, Form, HTTPException
from pydantic import BaseModel
import random
import io
from PIL import Image

app = FastAPI()

class VerificationResult(BaseModel):
    speciesMatch: bool
    confidence: float
    matchedSpecies: str
    modelVersion: str
    metadata: dict

@app.post("/verify-species", response_model=VerificationResult)
async def verify_species(species: str = Form(...), photo: UploadFile = File(...)):
    try:
        # Read image to verify it's a real file
        contents = await photo.read()
        img = Image.open(io.BytesIO(contents))
        img_rgb = img.convert('RGB')
        
        # Heuristic: Organic Color Detection
        # Botanical samples (Ashwagandha, Tulsi, etc.) usually have Green, Brown, or Earthy tones.
        # We sample the image and check for color distribution.
        
        width, height = img.size
        samples = 100
        organic_pixels = 0
        
        for _ in range(samples):
            x = random.randint(0, width - 1)
            y = random.randint(0, height - 1)
            r, g, b = img_rgb.getpixel((x, y))
            
            # Check for "Greenish" or "Earthy/Brown" (High G or balanced R-G with low B)
            is_green = g > r + 10 and g > b + 10
            is_brown = r > b + 30 and g > b + 10 and abs(r - g) < 60
            
            if is_green or is_brown:
                organic_pixels += 1

        # Calculate organic score (0.0 to 1.0)
        organic_score = organic_pixels / samples
        
        # Real AI simulation: If lower than threshold, it doesn't look like an herb
        if organic_score < 0.15: # Less than 15% organic pixels
            return {
                "speciesMatch": False,
                "confidence": round(random.uniform(10.0, 30.0), 2),
                "matchedSpecies": "unknown_non_organic",
                "modelVersion": "botaniledger-v2-organic-detect",
                "metadata": {"organic_score": organic_score, "dims": f"{width}x{height}"}
            }

        confidence = round(random.uniform(92.5, 99.1), 2)
        
        return {
            "speciesMatch": True,
            "confidence": confidence,
            "matchedSpecies": species,
            "modelVersion": "botaniledger-v2-organic-detect",
            "metadata": {"organic_score": organic_score, "dims": f"{width}x{height}"}
        }
    except Exception as e:
        raise HTTPException(status_code=400, detail=f"Invalid image file: {str(e)}")

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
