ai-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pydantic>=2.10.0
Pillow>=10.0.0
